Allow infinite scroll in heroes view controller

+++ Modules/Sources/HeroesFeature/HeroesView/HeroesEnvironment.swift
@@ -3,13 +3,13 @@ import ComposableArchitecture
 import Foundation
 
 public struct HeroesEnvironment {
-    var marvelCharactersLoader: () -> AnyPublisher<[MarvelCharacter], Error>
+    var marvelCharactersLoader: (Int) -> AnyPublisher<[MarvelCharacter], Error>
     var loadThumbnail: (URL) -> AnyPublisher<Data, Never>
     var loadMarvelCharacters: (String) -> AnyPublisher<[MarvelCharacter], Error>
     var mainQueue: AnySchedulerOf<DispatchQueue>
     
     public init(
-        marvelCharactersLoader : @escaping () -> AnyPublisher<[MarvelCharacter], Error>,
+        marvelCharactersLoader : @escaping (Int) -> AnyPublisher<[MarvelCharacter], Error>,
         loadThumbnail: @escaping (URL) -> AnyPublisher<Data, Never>,
         loadMarvelCharacters: @escaping (String) -> AnyPublisher<[MarvelCharacter], Error>,
         mainQueue: AnySchedulerOf<DispatchQueue>
